Add getLIC and renderLIC functions to generate a LIC texture and render it to the OpenGL widget

+++ Utilities/MyMatrix.h
@@ -25,31 +25,31 @@ class MyMatrix3x3;
 class MyMatrix2x2 {
 public:
     inline MyMatrix2x2();
-    inline MyMatrix2x2(double x);
+    inline MyMatrix2x2(float x);
     inline MyMatrix2x2(const MyMatrix2x2 &that);
 
-    inline MyMatrix2x2(double M00, double M01,
-                       double M10, double M11);
-    inline MyMatrix2x2(double M[2][2]);
+    inline MyMatrix2x2(float M00, float M01,
+                       float M10, float M11);
+    inline MyMatrix2x2(float M[2][2]);
 
-    inline MyMatrix2x2 &set      (const double d);
-    inline MyMatrix2x2 &operator=(const double d);
+    inline MyMatrix2x2 &set      (const float d);
+    inline MyMatrix2x2 &operator=(const float d);
 
     inline MyMatrix2x2 &set      (const MyMatrix2x2 &that);
     inline MyMatrix2x2 &operator=(const MyMatrix2x2 &that);
 
-    inline MyMatrix2x2 &set			 (double M[2][2]);
-    inline MyMatrix2x2 &operator=(double M[2][2]);
+    inline MyMatrix2x2 &set			 (float M[2][2]);
+    inline MyMatrix2x2 &operator=(float M[2][2]);
 
     inline int operator!=(const MyMatrix2x2 &that)const;
     inline int operator==(const MyMatrix2x2 &that)const;
 
-    inline int operator==(double d) const;
-    inline int operator!=(double d) const;
+    inline int operator==(float d) const;
+    inline int operator!=(float d) const;
 
-    inline MyMatrix2x2 &operator+=(double d);
-    inline MyMatrix2x2 &operator-=(double d);
-    inline MyMatrix2x2 &operator*=(double d);
+    inline MyMatrix2x2 &operator+=(float d);
+    inline MyMatrix2x2 &operator-=(float d);
+    inline MyMatrix2x2 &operator*=(float d);
 
     // component-wise operations.
     inline MyMatrix2x2 &operator+=(const MyMatrix2x2 &that);
@@ -64,13 +64,13 @@ public:
     inline MyMatrix2x2 &setIdentity     ();
 
 public:
-    double entry[2][2];
+    float entry[2][2];
 
 };
 
-inline MyMatrix2x2 operator+(const MyMatrix2x2 &a, double b);
-inline MyMatrix2x2 operator-(const MyMatrix2x2 &a, double b);
-inline MyMatrix2x2 operator*(const MyMatrix2x2 &a, double b);
+inline MyMatrix2x2 operator+(const MyMatrix2x2 &a, float b);
+inline MyMatrix2x2 operator-(const MyMatrix2x2 &a, float b);
+inline MyMatrix2x2 operator*(const MyMatrix2x2 &a, float b);
 
 inline MyMatrix2x2 operator+(const MyMatrix2x2 &a, const MyMatrix2x2 &b);
 inline MyMatrix2x2 operator-(const MyMatrix2x2 &a, const MyMatrix2x2 &b);
@@ -80,7 +80,7 @@ inline MyMatrix2x2 multiply(const MyMatrix2x2 &a, const MyMatrix2x2 &b);
 inline MyVector2d   operator*(const MyMatrix2x2 &a, const MyVector2d   &b);
 inline MyVector2d   operator*(const MyVector2d   &a, const MyMatrix2x2 &b);
 
-inline double determinant(const MyMatrix2x2 &a);
+inline float determinant(const MyMatrix2x2 &a);
 
 inline MyMatrix2x2 transpose(const MyMatrix2x2 &a);
 inline MyMatrix2x2   inverse(const MyMatrix2x2 &a);
@@ -92,15 +92,15 @@ inline MyMatrix2x2::MyMatrix2x2() {
     entry[1][1] = 1;
 }
 
-inline MyMatrix2x2::MyMatrix2x2(double x) {
+inline MyMatrix2x2::MyMatrix2x2(float x) {
     entry[0][0] = x;
     entry[0][1] = x;
     entry[1][0] = x;
     entry[1][1] = x;
 }
 
-inline MyMatrix2x2::MyMatrix2x2(double M00, double M01,
-                                double M10, double M11) {
+inline MyMatrix2x2::MyMatrix2x2(float M00, float M01,
+                                float M10, float M11) {
     entry[0][0] = M00;
     entry[0][1] = M01;
     entry[1][0] = M10;
@@ -114,11 +114,11 @@ inline MyMatrix2x2::MyMatrix2x2(const MyMatrix2x2 &that) {
     entry[1][1] = that.entry[1][1];
 };
 
-inline MyMatrix2x2 &MyMatrix2x2::set(const double d) {
+inline MyMatrix2x2 &MyMatrix2x2::set(const float d) {
     return (*this)=d;
 }
 
-inline MyMatrix2x2 &MyMatrix2x2::operator=(const double d) {
+inline MyMatrix2x2 &MyMatrix2x2::operator=(const float d) {
     entry[0][0] = d;
     entry[0][1] = d;
 
@@ -140,11 +140,11 @@ inline MyMatrix2x2 &MyMatrix2x2::operator=(const MyMatrix2x2 &that) {
     return (*this);
 };
 
-inline MyMatrix2x2 &MyMatrix2x2::set(double M[2][2]) {
+inline MyMatrix2x2 &MyMatrix2x2::set(float M[2][2]) {
     return (*this)=M;
 }
 
-inline MyMatrix2x2 &MyMatrix2x2::operator=(double M[2][2]) {
+inline MyMatrix2x2 &MyMatrix2x2::operator=(float M[2][2]) {
     entry[0][0] = M[0][0];
     entry[0][1] = M[0][1];
 
@@ -153,14 +153,14 @@ inline MyMatrix2x2 &MyMatrix2x2::operator=(double M[2][2]) {
     return (*this);
 };
 
-inline int MyMatrix2x2::operator==(double d) const {
+inline int MyMatrix2x2::operator==(float d) const {
     return  ( (entry[0][0] == d) &&
             (entry[0][1] == d) &&
             (entry[1][0] == d) &&
             (entry[1][1] == d) );
 }
 
-inline int MyMatrix2x2::operator!=(double d) const {
+inline int MyMatrix2x2::operator!=(float d) const {
     return  ( (entry[0][0] != d) ||
             (entry[0][1] != d) ||
             (entry[1][0] != d) ||
@@ -181,19 +181,19 @@ inline int MyMatrix2x2::operator!=(const MyMatrix2x2 &that)const {
             (entry[1][1] != that.entry[1][1]) );
 }
 
-inline MyMatrix2x2 &MyMatrix2x2::operator+=(double d) {
+inline MyMatrix2x2 &MyMatrix2x2::operator+=(float d) {
     entry[0][0] += d; entry[1][0] += d;
     entry[0][1] += d; entry[1][1] += d;
     return (*this);
 }
 
-inline MyMatrix2x2 &MyMatrix2x2::operator-=(double d) {
+inline MyMatrix2x2 &MyMatrix2x2::operator-=(float d) {
     entry[0][0] -= d; entry[1][0] -= d;
     entry[0][1] -= d; entry[1][1] -= d;
     return (*this);
 }
 
-inline MyMatrix2x2 &MyMatrix2x2::operator*=(double d) {
+inline MyMatrix2x2 &MyMatrix2x2::operator*=(float d) {
     entry[0][0] *= d; entry[1][0] *= d;
     entry[0][1] *= d; entry[1][1] *= d;
     return (*this);
@@ -218,11 +218,11 @@ inline MyMatrix2x2 &MyMatrix2x2::operator*=(const MyMatrix2x2 &that) {
     //entry[0][1] *= that.entry[0][1]; entry[1][1] *= that.entry[1][1];
 
     /*  Updated code */
-    double item00 = entry[0][0] * that.entry[0][0] + entry[0][1] * that.entry[1][0];
-    double item01 = entry[0][0] * that.entry[0][1] + entry[0][1] * that.entry[1][1];
+    float item00 = entry[0][0] * that.entry[0][0] + entry[0][1] * that.entry[1][0];
+    float item01 = entry[0][0] * that.entry[0][1] + entry[0][1] * that.entry[1][1];
 
-    double item10 = entry[1][0] * that.entry[0][0] + entry[1][1] * that.entry[1][0];
-    double item11 = entry[1][0] * that.entry[0][1] + entry[1][1] * that.entry[1][1];
+    float item10 = entry[1][0] * that.entry[0][0] + entry[1][1] * that.entry[1][0];
+    float item11 = entry[1][0] * that.entry[0][1] + entry[1][1] * that.entry[1][1];
     entry[0][0] = item00;
     entry[0][1] = item01;
     entry[1][0] = item10;
@@ -256,28 +256,28 @@ inline MyMatrix2x2 &MyMatrix2x2::setIdentity() {
     return (*this);
 };
 
-inline MyMatrix2x2 operator+(const MyMatrix2x2 &a,double b) {
+inline MyMatrix2x2 operator+(const MyMatrix2x2 &a,float b) {
     return (MyMatrix2x2(a)+=b);
 }
 
-inline MyMatrix2x2 operator-(const MyMatrix2x2 &a,double b) {
+inline MyMatrix2x2 operator-(const MyMatrix2x2 &a,float b) {
     return (MyMatrix2x2(a)-=b);
 }
 
-inline MyMatrix2x2 operator*(const MyMatrix2x2 &a,double b) {
+inline MyMatrix2x2 operator*(const MyMatrix2x2 &a,float b) {
     return (MyMatrix2x2(a)*=b);
 }
 
-inline MyMatrix2x2 operator+(double a, const MyMatrix2x2 &b) {
+inline MyMatrix2x2 operator+(float a, const MyMatrix2x2 &b) {
     return b+a;
 }
 
-inline MyMatrix2x2 operator-(double a, const MyMatrix2x2 &b) {
+inline MyMatrix2x2 operator-(float a, const MyMatrix2x2 &b) {
     return MyMatrix2x2(a-b.entry[0][0],a-b.entry[0][1],
             a-b.entry[1][0],a-b.entry[1][1]);
 }
 
-inline MyMatrix2x2 operator*(double a, const MyMatrix2x2 &b) {
+inline MyMatrix2x2 operator*(float a, const MyMatrix2x2 &b) {
     return b*a;
 }
 
@@ -309,7 +309,7 @@ inline MyVector2d operator*(const MyVector2d &a,const MyMatrix2x2 &b) {
             a.entry[0]*b.entry[0][1] + a.entry[1]*b.entry[1][1]);
 }
 
-inline double determinant(const MyMatrix2x2 &a) {
+inline float determinant(const MyMatrix2x2 &a) {
     return ( a.entry[0][0] * a.entry[1][1] - a.entry[0][1] * a.entry[1][0] );
 }
 
@@ -323,7 +323,7 @@ inline MyMatrix2x2 transpose(const MyMatrix2x2 &a) {
 
 inline MyMatrix2x2 inverse(const MyMatrix2x2 &a) {
     MyMatrix2x2 tmp;
-    double dmt;
+    float dmt;
 
     if ((dmt=determinant(a))!= 0.0) {
         tmp.entry[0][0] = a.entry[1][1]/dmt;
@@ -340,37 +340,37 @@ inline MyMatrix2x2 inverse(const MyMatrix2x2 &a) {
 class MyMatrix3x3 {
 public:
     inline MyMatrix3x3();
-    inline MyMatrix3x3(double x);
+    inline MyMatrix3x3(float x);
     inline MyMatrix3x3(const MyMatrix3x3 &that);
     inline MyMatrix3x3(const MyVector3d &v1, const MyVector3d &v2, const MyVector3d &v3);
 
-    inline MyMatrix3x3(double M00, double M01, double M02,
-                       double M10, double M11, double M12,
-                       double M20, double M21, double M22);
-    inline MyMatrix3x3(double M[3][3]);
+    inline MyMatrix3x3(float M00, float M01, float M02,
+                       float M10, float M11, float M12,
+                       float M20, float M21, float M22);
+    inline MyMatrix3x3(float M[3][3]);
 
-    inline MyMatrix3x3 &set      (const double d);
-    inline MyMatrix3x3 &operator=(const double d);
+    inline MyMatrix3x3 &set      (const float d);
+    inline MyMatrix3x3 &operator=(const float d);
 
     inline MyMatrix3x3 &set      (const MyMatrix3x3 &that);
     inline MyMatrix3x3 &operator=(const MyMatrix3x3 &that);
 
-    inline MyMatrix3x3 &set			 (double M[3][3]);
-    inline MyMatrix3x3 &operator=(double M[3][3]);
+    inline MyMatrix3x3 &set			 (float M[3][3]);
+    inline MyMatrix3x3 &operator=(float M[3][3]);
 
     inline MyMatrix3x3 &set     (const MyVector3d &v1, const MyVector3d &v2, const MyVector3d &v3);
-    inline MyMatrix3x3 &set			(double M00, double M01, double M02,
-                                     double M10, double M11, double M12,
-                                     double M20, double M21, double M22);
+    inline MyMatrix3x3 &set			(float M00, float M01, float M02,
+                                     float M10, float M11, float M12,
+                                     float M20, float M21, float M22);
     inline int operator!=(const MyMatrix3x3 &that)const;
     inline int operator==(const MyMatrix3x3 &that)const;
 
-    inline int operator==(double d) const;
-    inline int operator!=(double d) const;
+    inline int operator==(float d) const;
+    inline int operator!=(float d) const;
 
-    inline MyMatrix3x3 &operator+=(double d);
-    inline MyMatrix3x3 &operator-=(double d);
-    inline MyMatrix3x3 &operator*=(double d);
+    inline MyMatrix3x3 &operator+=(float d);
+    inline MyMatrix3x3 &operator-=(float d);
+    inline MyMatrix3x3 &operator*=(float d);
 
     // component-wise operations.
     inline MyMatrix3x3 &operator+=(const MyMatrix3x3 &that);
@@ -385,13 +385,13 @@ public:
     inline MyMatrix3x3 &setIdentity     ();
 
 public:
-    double entry[3][3];
+    float entry[3][3];
 
 };
 
-inline MyMatrix3x3 operator+(const MyMatrix3x3 &a, double b);
-inline MyMatrix3x3 operator-(const MyMatrix3x3 &a, double b);
-inline MyMatrix3x3 operator*(const MyMatrix3x3 &a, double b);
+inline MyMatrix3x3 operator+(const MyMatrix3x3 &a, float b);
+inline MyMatrix3x3 operator-(const MyMatrix3x3 &a, float b);
+inline MyMatrix3x3 operator*(const MyMatrix3x3 &a, float b);
 
 inline MyMatrix3x3 operator+(const MyMatrix3x3 &a, const MyMatrix3x3 &b);
 inline MyMatrix3x3 operator-(const MyMatrix3x3 &a, const MyMatrix3x3 &b);
@@ -403,7 +403,7 @@ inline MyMatrix3x3 othoconjugate(const MyMatrix3x3 &a, const MyMatrix3x3 &b);
 inline MyVector3d   operator*(const MyMatrix3x3 &a, const MyVector3d   &b);
 inline MyVector3d   operator*(const MyVector3d   &a, const MyMatrix3x3 &b);
 
-inline double determinant(const MyMatrix3x3 &a);
+inline float determinant(const MyMatrix3x3 &a);
 
 inline MyMatrix3x3 transpose(const MyMatrix3x3 &a);
 inline MyMatrix3x3   inverse(const MyMatrix3x3 &a);
@@ -420,7 +420,7 @@ inline MyMatrix3x3::MyMatrix3x3() {
     entry[2][2] = 1;
 }
 
-inline MyMatrix3x3::MyMatrix3x3(double x) {
+inline MyMatrix3x3::MyMatrix3x3(float x) {
     entry[0][0] = x;
     entry[0][1] = x;
     entry[0][2] = x;
@@ -432,9 +432,9 @@ inline MyMatrix3x3::MyMatrix3x3(double x) {
     entry[2][2] = x;
 }
 
-inline MyMatrix3x3::MyMatrix3x3(double M00, double M01, double M02,
-                                double M10, double M11, double M12,
-                                double M20, double M21, double M22) {
+inline MyMatrix3x3::MyMatrix3x3(float M00, float M01, float M02,
+                                float M10, float M11, float M12,
+                                float M20, float M21, float M22) {
     entry[0][0] = M00;
     entry[0][1] = M01;
     entry[0][2] = M02;
@@ -470,11 +470,11 @@ inline MyMatrix3x3::MyMatrix3x3(const MyVector3d &v1, const MyVector3d &v2, cons
     entry[2][2] = v3.entry[2];
 }
 
-inline MyMatrix3x3 &MyMatrix3x3::set(const double d) {
+inline MyMatrix3x3 &MyMatrix3x3::set(const float d) {
     return (*this)=d;
 }
 
-inline MyMatrix3x3 &MyMatrix3x3::operator=(const double d) {
+inline MyMatrix3x3 &MyMatrix3x3::operator=(const float d) {
     entry[0][0] = d;
     entry[0][1] = d;
     entry[0][2] = d;
@@ -507,11 +507,11 @@ inline MyMatrix3x3 &MyMatrix3x3::operator=(const MyMatrix3x3 &that) {
     return (*this);
 };
 
-inline MyMatrix3x3 &MyMatrix3x3::set(double M[3][3]) {
+inline MyMatrix3x3 &MyMatrix3x3::set(float M[3][3]) {
     return (*this)=M;
 }
 
-inline MyMatrix3x3 &MyMatrix3x3::operator=(double M[3][3]) {
+inline MyMatrix3x3 &MyMatrix3x3::operator=(float M[3][3]) {
     entry[0][0] = M[0][0];
     entry[0][1] = M[0][1];
     entry[0][2] = M[0][2];
@@ -539,9 +539,9 @@ inline MyMatrix3x3 &MyMatrix3x3::set(const MyVector3d &v1, const MyVector3d &v2,
     return (*this);
 }
 
-inline MyMatrix3x3 &MyMatrix3x3::set			(double M00, double M01, double M02,
-                                                 double M10, double M11, double M12,
-                                                 double M20, double M21, double M22)
+inline MyMatrix3x3 &MyMatrix3x3::set			(float M00, float M01, float M02,
+                                                 float M10, float M11, float M12,
+                                                 float M20, float M21, float M22)
 {
     entry[0][0] = M00;
     entry[0][1] = M01;
@@ -555,13 +555,13 @@ inline MyMatrix3x3 &MyMatrix3x3::set			(double M00, double M01, double M02,
     return (*this);
 }
 
-inline int MyMatrix3x3::operator==(double d) const {
+inline int MyMatrix3x3::operator==(float d) const {
     return  ( (entry[0][0] == d) && (entry[0][1] == d) && (entry[0][2] == d) &&
             (entry[1][0] == d) && (entry[1][1] == d) && (entry[1][2] == d) &&
             (entry[2][0] == d) && (entry[2][1] == d) && (entry[2][2] == d));
 }
 
-inline int MyMatrix3x3::operator!=(double d) const {
+inline int MyMatrix3x3::operator!=(float d) const {
     return  ( (entry[0][0] != d) || (entry[0][1] != d) || (entry[0][2] != d) ||
             (entry[1][0] != d) || (entry[1][1] != d) || (entry[1][2] != d) ||
             (entry[2][0] != d) || (entry[2][1] != d) || (entry[2][2] != d));
@@ -579,21 +579,21 @@ inline int MyMatrix3x3::operator!=(const MyMatrix3x3 &that)const {
             (entry[2][0] != that.entry[2][0]) || (entry[2][1] != that.entry[2][1]) || (entry[2][2] != that.entry[2][2]));
 }
 
-inline MyMatrix3x3 &MyMatrix3x3::operator+=(double d) {
+inline MyMatrix3x3 &MyMatrix3x3::operator+=(float d) {
     entry[0][0] += d; entry[0][1] += d; entry[0][2] += d;
     entry[1][0] += d; entry[1][1] += d; entry[1][2] += d;
     entry[2][0] += d; entry[2][1] += d; entry[2][2] += d;
     return (*this);
 }
 
-inline MyMatrix3x3 &MyMatrix3x3::operator-=(double d) {
+inline MyMatrix3x3 &MyMatrix3x3::operator-=(float d) {
     entry[0][0] -= d; entry[0][1] -= d; entry[0][2] -= d;
     entry[1][0] -= d; entry[1][1] -= d; entry[1][2] -= d;
     entry[2][0] -= d; entry[2][1] -= d; entry[2][2] -= d;
     return (*this);
 }
 
-inline MyMatrix3x3 &MyMatrix3x3::operator*=(double d) {
+inline MyMatrix3x3 &MyMatrix3x3::operator*=(float d) {
     entry[0][0] *= d; entry[0][1] *= d; entry[0][2] *= d;
     entry[1][0] *= d; entry[1][1] *= d; entry[1][2] *= d;
     entry[2][0] *= d; entry[2][1] *= d; entry[2][2] *= d;
@@ -666,29 +666,29 @@ inline MyMatrix3x3 &MyMatrix3x3::setIdentity() {
     return (*this);
 };
 
-inline MyMatrix3x3 operator+(const MyMatrix3x3 &a,double b) {
+inline MyMatrix3x3 operator+(const MyMatrix3x3 &a,float b) {
     return (MyMatrix3x3(a)+=b);
 }
 
-inline MyMatrix3x3 operator-(const MyMatrix3x3 &a,double b) {
+inline MyMatrix3x3 operator-(const MyMatrix3x3 &a,float b) {
     return (MyMatrix3x3(a)-=b);
 }
 
-inline MyMatrix3x3 operator*(const MyMatrix3x3 &a,double b) {
+inline MyMatrix3x3 operator*(const MyMatrix3x3 &a,float b) {
     return (MyMatrix3x3(a)*=b);
 }
 
-inline MyMatrix3x3 operator+(double a, const MyMatrix3x3 &b) {
+inline MyMatrix3x3 operator+(float a, const MyMatrix3x3 &b) {
     return b+a;
 }
 
-inline MyMatrix3x3 operator-(double a, const MyMatrix3x3 &b) {
+inline MyMatrix3x3 operator-(float a, const MyMatrix3x3 &b) {
     return MyMatrix3x3(a-b.entry[0][0],a-b.entry[0][1],a-b.entry[0][2],
             a-b.entry[1][0],a-b.entry[1][1],a-b.entry[1][2],
             a-b.entry[2][0],a-b.entry[2][1],a-b.entry[2][2]);
 }
 
-inline MyMatrix3x3 operator*(double a, const MyMatrix3x3 &b) {
+inline MyMatrix3x3 operator*(float a, const MyMatrix3x3 &b) {
     return b*a;
 }
 
@@ -738,7 +738,7 @@ inline MyVector3d operator*(const MyVector3d &a,const MyMatrix3x3 &b) {
             a.entry[0]*b.entry[0][2] + a.entry[1]*b.entry[1][2] + a.entry[2]*b.entry[2][2]);
 }
 
-inline double determinant(const MyMatrix3x3 &a) {
+inline float determinant(const MyMatrix3x3 &a) {
     return ( a.entry[0][0] * a.entry[1][1] * a.entry[2][2] - a.entry[2][0] * a.entry[1][1] * a.entry[0][2]
             + a.entry[1][0] * a.entry[2][1] * a.entry[0][2] - a.entry[0][0] * a.entry[2][1] * a.entry[1][2]
             + a.entry[2][0] * a.entry[0][1] * a.entry[1][2] - a.entry[1][0] * a.entry[0][1] * a.entry[2][2]);
@@ -760,7 +760,7 @@ inline MyMatrix3x3 transpose(const MyMatrix3x3 &a) {
 
 inline MyMatrix3x3 inverse(const MyMatrix3x3 &a) {
     MyMatrix3x3 tmp;
-    double dmt;
+    float dmt;
 
     if ((dmt=determinant(a))!= 0.0) {
         tmp.entry[0][0] = (a.entry[1][1] * a.entry[2][2] - a.entry[2][1] * a.entry[1][2])/dmt;
